requirement.txt
langchain
langchain_community
python-dotenv
openai
pypdf
faiss-gpu
streamlit